refactor: consolidate ANIMA prompt package
