+++ docs/architecture/python-backend-execution-roadmap.md
@@ -33,7 +33,7 @@ merged PR, the owning issue's evidence record, and every stated exit gate.
 | A - baseline | Complete; #191 is closed | Keep fixtures and analyzers current during later moves |
 | B - `nodes.py` extraction | Complete in B-11d / PR #356 | Preserve the audited compatibility shim until ADR-002 retirement gates are met |
 | C - feature contracts/behavior | Partially complete through S167-01a / PR #344 | Continue #167 and #169 in separate Contract/Move/Behavior PRs |
-| D - root consolidation | D-01, D-08, D-09, D-10, and D-12 complete on `dev`; D-11a/#385 and D-11b/#386 validated | Continue D-13; finish D-11 classification only after D-13 removes its root-package dependency |
+| D - root consolidation | D-01, D-08, D-09, D-10, D-12, and D-13 complete or in promotion; D-11a/#385 and D-11b/#386 validated | Finish the D-11 classification shim slice |
 | E - runtime ownership | Partial: E-02a and E-07a/E-07b integrated | Continue #187 only where canonical feature owners and explicit contracts exist |
 | F - typed boundaries | Partial patterns exist | Extend typed request/result/config and pure migration patterns feature by feature |
 | G - quality ratchet | G-01, G-02a/G-02b, and G-03a complete | Extend G-03 enrollment, then continue with G-04 through G-06 |
@@ -408,7 +408,7 @@ mechanical retirement series.
 | 15 | S167 backend seed reservation series | S167-01 through S167-03d COMPLETE on `dev`; S167-03e AiO cutover VALIDATED with isolated API/module/browser-load parity | Contract then Move then Behavior | #167 | Canonical AiO/node seams |
 | 16 | A169 stage pipeline series | A169-01 through A169-08 MERGED; A169-09 final adapter/integration VALIDATED in PR #372 | Contract then Behavior | #169 | Typed config and mechanical AiO move |
 | 17 | A169 first-pass cache policy | COMPLETE through CACHE-06; 4K/batch evidence VALIDATED in PR #380 | Contract then Behavior | #169 | Mechanical cache move and stable stage seam |
-| 18 | D-series canonical root consolidation | D-01 translation, D-08 filesystem, D-09 settings, D-10 profiles, and D-12 wildcard COMPLETE on `dev`; D-11a/#385 and D-11b/#386 autocomplete VALIDATED | Contract then Move | #186 | Phase B exit; per-feature behavior stable |
+| 18 | D-series canonical root consolidation | D-01 translation, D-08 filesystem, D-09 settings, D-10 profiles, D-12 wildcard, and D-13 ANIMA prompt COMPLETE or in promotion; D-11a/#385 and D-11b/#386 autocomplete VALIDATED | Finish the D-11 classification shim slice | #186 | Phase B exit; per-feature behavior stable |
 | 19 | E-series RuntimeServices/lifecycle | BLOCKED by canonical owners | Move/Contract, split PRs | #187 | Relevant D moves |
 | 20 | G-04 through G-06 and H | INCREMENTAL/LATER | Gate/Contract | #188 | Appropriate package and release evidence |
 
@@ -1378,6 +1378,18 @@ construction, stage ordering, depth/repeated-state control, and result assembly
 to the canonical expansion owner with an immutable snapshot input. Root keeps
 the public facade, selector construction, root resolution, budget preparation,
 and snapshot lifecycle; this completes D-12 without starting E-06.
+
+D-13 moves the dependency-light ANIMA prompt parser, normalization, ordering,
+knowledge, models, and correction core together to
+`easyuse_anima.prompt.anima`. The seven root `anima_prompt` modules remain
+explicit identity shims, and the Registry analyzer treats each directly
+importable shim as an external compatibility entry so packed closure remains
+complete. Internal Prompt Corrector, Prompt Builder/field, autocomplete
+classification, and root node consumers import only the canonical owner.
+`PACKAGE_DATA_DIR` still resolves to the repository `__easyuse_anima__`
+directory. Keeping the small, tightly coupled package in one Move avoids
+temporary cross-root dependencies; parse, normalize, order, correction, and
+classification behavior is unchanged.
 
 ## 12. Phase E — Runtime ownership and lifecycle
