+++ easyuse_anima/prompt/anima/__init__.py
@@ -1,0 +1,26 @@
+"""ANIMA prompt correction helpers.
+
+This package is intentionally dependency-light. It does not import torch,
+ComfyUI, model code, or taggers, so the same core can be reused by CLI tools
+and future UI nodes.
+"""
+
+from .correction import correct_prompt, inspect_prompt
+from .knowledge import (
+    KnowledgeBaseNotFound,
+    PromptKnowledgeBase,
+    load_knowledge_base,
+)
+from .models import CorrectionResult, ParsedPrompt, TagInfo, TagToken
+
+__all__ = [
+    "CorrectionResult",
+    "KnowledgeBaseNotFound",
+    "ParsedPrompt",
+    "PromptKnowledgeBase",
+    "TagInfo",
+    "TagToken",
+    "correct_prompt",
+    "inspect_prompt",
+    "load_knowledge_base",
+]
