+++ tests/fixtures/python_compatibility_surface.v1.json
@@ -118,8 +118,8 @@
   "expected_counts": {
     "root_entrypoints": 3,
     "excluded_preamble_implementation_bindings": 0,
-    "nodes_canonical_bindings": 294,
-    "nodes_legacy_bindings": 22,
+    "nodes_canonical_bindings": 297,
+    "nodes_legacy_bindings": 19,
     "nodes_public_exports": 18,
     "mapped_public_classes": 18,
     "unmapped_classes": 2,
@@ -2422,6 +2422,65 @@
       "lifecycle_state": "unsupported"
     },
     {
+      "id": "nodes_canonical_bindings:easyuse_anima.prompt.anima.parser:unsupported_test_only",
+      "surface": "nodes_canonical_bindings",
+      "symbols": {
+        "parse_prompt": "parse_prompt"
+      },
+      "classification": "unsupported_test_only",
+      "current_target": "nodes.py",
+      "canonical_target": "easyuse_anima.prompt.anima.parser",
+      "target_state": "canonical",
+      "identity_requirement": "not_applicable",
+      "binding_shape": "direct_import",
+      "import_target": "easyuse_anima.prompt.anima.parser",
+      "owner": "#184/#188",
+      "first_release": null,
+      "known_consumers": [
+        "repository tests may import this root name"
+      ],
+      "evidence": [
+        "AST:no nodes.py runtime load",
+        "test-only consumption is not public support evidence"
+      ],
+      "removal_gates": [
+        "test imports migrate to canonical targets",
+        "no production consumer evidence",
+        "separate B-10b removal review"
+      ],
+      "lifecycle_state": "unsupported"
+    },
+    {
+      "id": "nodes_canonical_bindings:easyuse_anima.prompt.anima:unsupported_test_only",
+      "surface": "nodes_canonical_bindings",
+      "symbols": {
+        "correct_prompt": "correct_prompt",
+        "load_knowledge_base": "load_knowledge_base"
+      },
+      "classification": "unsupported_test_only",
+      "current_target": "nodes.py",
+      "canonical_target": "easyuse_anima.prompt.anima",
+      "target_state": "canonical",
+      "identity_requirement": "not_applicable",
+      "binding_shape": "direct_import",
+      "import_target": "easyuse_anima.prompt.anima",
+      "owner": "#184/#188",
+      "first_release": null,
+      "known_consumers": [
+        "repository tests may import this root name"
+      ],
+      "evidence": [
+        "AST:no nodes.py runtime load",
+        "test-only consumption is not public support evidence"
+      ],
+      "removal_gates": [
+        "test imports migrate to canonical targets",
+        "no production consumer evidence",
+        "separate B-10b removal review"
+      ],
+      "lifecycle_state": "unsupported"
+    },
+    {
       "id": "nodes_canonical_bindings:easyuse_anima.prompt.artist_mix:unsupported_test_only",
       "surface": "nodes_canonical_bindings",
       "symbols": {
@@ -2787,65 +2846,6 @@
       "binding_shape": "direct_import",
       "import_target": "easyuse_anima.workflow",
       "owner": "#184/#188",
-      "first_release": null,
-      "known_consumers": [
-        "repository tests may import this root name"
-      ],
-      "evidence": [
-        "AST:no nodes.py runtime load",
-        "test-only consumption is not public support evidence"
-      ],
-      "removal_gates": [
-        "test imports migrate to canonical targets",
-        "no production consumer evidence",
-        "separate B-10b removal review"
-      ],
-      "lifecycle_state": "unsupported"
-    },
-    {
-      "id": "nodes_legacy_bindings:anima_prompt.parser:unsupported_test_only",
-      "surface": "nodes_legacy_bindings",
-      "symbols": {
-        "parse_prompt": "parse_prompt"
-      },
-      "classification": "unsupported_test_only",
-      "current_target": "nodes.py",
-      "canonical_target": null,
-      "target_state": "legacy_owner",
-      "identity_requirement": "not_applicable",
-      "binding_shape": "direct_import",
-      "import_target": "anima_prompt.parser",
-      "owner": "#184/#186",
-      "first_release": null,
-      "known_consumers": [
-        "repository tests may import this root name"
-      ],
-      "evidence": [
-        "AST:no nodes.py runtime load",
-        "test-only consumption is not public support evidence"
-      ],
-      "removal_gates": [
-        "test imports migrate to canonical targets",
-        "no production consumer evidence",
-        "separate B-10b removal review"
-      ],
-      "lifecycle_state": "unsupported"
-    },
-    {
-      "id": "nodes_legacy_bindings:anima_prompt:unsupported_test_only",
-      "surface": "nodes_legacy_bindings",
-      "symbols": {
-        "correct_prompt": "correct_prompt",
-        "load_knowledge_base": "load_knowledge_base"
-      },
-      "classification": "unsupported_test_only",
-      "current_target": "nodes.py",
-      "canonical_target": null,
-      "target_state": "legacy_owner",
-      "identity_requirement": "not_applicable",
-      "binding_shape": "direct_import",
-      "import_target": "anima_prompt",
-      "owner": "#184/#186",
       "first_release": null,
       "known_consumers": [
         "repository tests may import this root name"
